com#4 added new menu tiles & upgraded gradle8.10.2
