+++ Aerocart/android/app/build.gradle
@@ -29,11 +29,11 @@ if (flutterVersionName == null) {
 android {
     namespace = "com.example.amazon_clone"
     compileSdk = 34
-    ndkVersion = flutter.ndkVersion
+    ndkVersion =  "27.0.12077973"
 
     compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
+        sourceCompatibility = JavaVersion.VERSION_21
+        targetCompatibility = JavaVersion.VERSION_21
     }
 
     defaultConfig {
