+++ Aerocart/lib/layout/screen_layout.dart
@@ -3,7 +3,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:amazon_clone/pages/cart_page.dart';
 import 'package:amazon_clone/pages/home_pge.dart';
-import 'package:amazon_clone/pages/user_page.dart';
+import 'package:amazon_clone/pages/menu.dart';
 import 'package:provider/provider.dart';
 
 import '../provider/user_details_provider.dart';
@@ -20,7 +20,7 @@ class _ScreenLayoutState extends State<ScreenLayout> {
   final List<Widget> tablist = [
     const HomePage(),
     const CartPage(),
-    const UserPage(),
+    const Menu(),
   ];
 
   Stream<int> getCartItemCountStream() {
@@ -73,8 +73,8 @@ class _ScreenLayoutState extends State<ScreenLayout> {
                         label: "Cart",
                       ),
                       const BottomNavigationBarItem(
-                        icon: Icon(Icons.account_circle_outlined),
-                        label: "User",
+                        icon: Icon(Icons.menu),
+                        label: "More",
                       ),
                     ],
                   );
